feat(kernel): enhance ARM64 type definitions and error handling
This update introduces several improvements across ARM64 kernel files. Notably, it adds the `EOPNOTSUPP` error code definition in `errno.h`, updates type definitions in `types.h` to ensure compatibility with LLP64 architectures, and modifies various source files to use `uintptr_t` for pointer arithmetic. Additionally, it enforces size and alignment checks for the `neverc_krt_dir_filter_scope` structure in `nvk_dir.h`. These changes enhance type safety and error handling throughout the kernel codebase.

+++ runtime/android/kernel/arm64/include/linux/sched.h
@@ -11,7 +11,7 @@ struct pid;         /* opaque */
 
 static __always_inline struct task_struct *get_current(void)
 {
-	unsigned long sp_el0;
+	uintptr_t sp_el0;
 	__asm__ __volatile__("mrs %0, sp_el0" : "=r"(sp_el0));
 	return (struct task_struct *)sp_el0;
 }
